Consolidate server for easier backend communication
