+++ package.json
@@ -4,7 +4,7 @@
   "main": "server.js",
   "type": "module",
   "scripts": {
-    "start": "node data.js && node server.js",
+    "start": "node data.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
