--- server.js
@@ -1,59 +1,0 @@
-
-import express from 'express';
-const app = express();
-import cors from 'cors';
-const corsOptions ={
-  origin: true, 
-  credentials:true,            //access-control-allow-credentials:true
-  optionSuccessStatus:200,
-}
-app.use(cors(corsOptions));
-import { application, response } from 'express';
-
-//Node-fetch
-import fetch from 'node-fetch';
-
-//puppeteer dependencies
-import start from './data.js'
-
-app.use(express.json())
-
-app.set('port', process.env.PORT || 8080);
-app.locals.title = 'Project Megaphone Server'
-app.locals.text = {email: null};
-
-const postText = (text) => {
-  fetch('/megaphone-ai-api.herokuapp.com/api/v1/extractions', 
-  {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json'},
-    body: JSON.stringify(text)
-  })
-  .then(response => response.json())
-  .then(data => console.log("what we get from posting", data))
-  .catch(error => console.log(error))
-}
-
-const getText = async (email) => {
-  let text = await start("http://www." + email.body.hd);
-  app.locals.text = { data: text }
-  console.log("the text", text)
-  postText(text)
-}
-
-app.get('/', (req, res) => {
-  res.send(app.locals.text)
-})
-
-
-app.post('/', (req, res) => {
-  response.setHeader('Content-Type', 'application/json')
-  res.send(app.locals.text)
-  console.log(req)
-  getText(req)
-  return res.send.json( req )
-})
-
-app.listen(app.get('port'), () => {
-  console.log(`${app.locals.title} is now running on ${app.get('port')}!`)
-})
